move platform independent code from MainWindow.xaml.cs to Core.cs
