+++ MapMaker/MapMaker/MapMaker.csproj
@@ -56,6 +56,7 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
+    <Compile Include="Classes\Core.cs" />
     <Compile Include="Classes\Delta.cs" />
     <Compile Include="Classes\ProjectSettings.cs" />
     <Compile Include="Classes\SHP\SHPElement.cs" />
